requirements.txt
anthropic==0.100.0
sentence-transformers==3.3.1
snowflake-connector-python==3.12.3
pymupdf==1.24.13
pdfplumber==0.11.4
tiktoken==0.8.0
python-dotenv==1.0.1
streamlit==1.40.1
pydantic==2.9.2
pytest==8.3.3
